+++ cloudsplaining/output/triage_worksheet.py
@@ -14,14 +14,16 @@ def create_triage_worksheet(account_name, results, output_directory):
     Create a triage spreadsheet for account owners to fill out. They can specify whether it is a false positive
     depending on context or whether or not it needs to be fixed.
     """
-    triage_spreadsheet_file = os.path.join(output_directory, f"iam-triage-{account_name}.csv")
+    triage_spreadsheet_file = os.path.join(
+        output_directory, f"iam-triage-{account_name}.csv"
+    )
     csv_fieldnames = [
         "PolicyName",
         "Type",
         "ManagedBy",
         "Services",  # ServicesCount
         "Actions",  # ActionsCount
-        "Justification"
+        "Justification",
     ]
     if os.path.exists(triage_spreadsheet_file):
         os.remove(triage_spreadsheet_file)