+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
-## 0.0.3 (Unreleased)
+## 0.0.3 (2020-05-01)
 * Provide option to skip opening HTML report (`--skip-open-report`)
+* Provide report indicator on whether it is assumable by compute services
+* Dropdown menu for report
 
 ## 0.0.2 (2020-04-30)
 * Quick markdown bug fix